mboxd_msg: Fix overflow in get_suggested_timeout()
The suggested timeout is a two-byte unsigned value, but when introducing the
debug message it was stored in a one-byte unsigned value. This can lead
to overflow under some circumstances.

Fixes: 285195987721 ("mboxd: Introduce a new DEBUG log level")

+++ mboxd_msg.c
@@ -158,7 +158,7 @@ static uint16_t get_suggested_timeout(struct mbox_context *context)
 {
 	struct window_context *window = find_largest_window(context);
 	uint32_t max_size_mb = window ? (window->size >> 20) : 0;
-	uint8_t ret;
+	uint16_t ret;
 
 	ret = align_up(max_size_mb * FLASH_ACCESS_MS_PER_MB, 1000) / 1000;
